backend/production_requirements.txt
# Production FastAPI Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# Production ASGI server
uvloop>=0.19.0  # Better event loop for production

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.13.0  # Database migrations
psycopg2-binary>=2.9.0  # PostgreSQL adapter for production

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0

# Machine Learning and AI
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
sentence-transformers>=2.2.0

# Object Detection
ultralytics>=8.0.0
opencv-python>=4.8.0

# Text Processing and NER
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0

# File handling and utilities
requests>=2.31.0
aiofiles>=23.0.0
python-magic>=0.4.27

# Production Monitoring & Logging
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0
prometheus-client>=0.19.0
psutil>=5.9.0

# Caching (Production)
redis>=5.0.0
python-redis-lock>=4.0.0

# Background Tasks
celery>=5.3.0
kombu>=5.3.0

# HTTP Client
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.10.0

# Image Processing & Annotation
imageio>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Email Support
emails>=0.6.0
jinja2>=3.1.0

# Configuration Management
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.8.0

# Validation
email-validator>=2.0.0
python-dateutil>=2.8.0

# File Upload Security
python-magic>=0.4.27
defusedxml>=0.7.1

# Production Database Backup
pg_dump>=1.0.0

# SSL/TLS Support
certifi>=2023.0.0

# Production Performance
orjson>=3.9.0  # Faster JSON serialization

# Resource Management
resource>=0.2.1

# Process Management
supervisor>=4.2.0

# Health Checks
health-check>=1.3.3

# API Documentation (Production)
python-markdown>=3.5.0

# Environment Management
python-decouple>=3.8.0

# Timezone Support
pytz>=2023.3

# UUID Generation
uuid>=1.30

# Production Logging
python-json-logger>=2.0.0
loguru>=0.7.0

# Rate Limiting
slowapi>=0.1.9

# Input Validation
validators>=0.22.0

# File Type Detection
python-magic-bin>=0.4.14  # For Windows compatibility

# Production Database Pool
SQLAlchemy-Utils>=0.41.0 